ta la

+++ Mecanico.h
@@ -3,7 +3,7 @@
 struct Mecanico {
 	string nome = "";
 	string marca = "";
-	float preco_reparacao_por_dia = 0;
+	int preco_reparacao_por_dia = 0;
 };
 
 Mecanico CriarMecanico(LinhasFicheiro& marcas);
